requirements.txt
numpy>=1.20.0
biopython>=1.79
pytest>=6.0.0